Add files via upload
[Quick Data Processing]
In phase 1 we utilize and create a quick data processing pipeline utilizing methods from the original project such as the following : 
Anomaly logger in order to detect anomalies within the datasets such as missing, corrupted or malformed metadata fields that might cause challenges in the machine learning model, we can call this "bad data" in which we do not use data that is not helpful for detecting censored events.
Fixing metadata fields such as large, ambiguous and very challenging to read for analysis, so we can utilize standardizing timestamps in order to make timestamps for eligible for data processing. 
Minibatch approach utilization, we can utilize the minibatch approach strategy in order to get a small sample of data from this large dataset so data processing and analysis can be more efficient and quick. 
Finally we can save the cleaned data once data is cleaned through the anomaly logger and fixed we can then save the data into a separate file that we can use for data processing. 

User input implementation : Allow the user to quickly input how much data to get from the dataset and which countries(s) we are selecting the data from, this is helpful for strictly focusing on a portion of data.

Further Improvements : We can began to expand this pipeline further by implementing more anomaly flags in order to detect corrupted data, we improve upon the data fixing as well such as making certain metadata fields much simpler and or adding more custom metadata fields flags as well,  we can incorporate the measurement score approach for phase 2 machine learning implementation.